use NIF template methods and reorg into NIF module
