+++ src/exla/exla.cc
@@ -16,20 +16,15 @@ ErlNifResourceType *OP_RES_TYPE, *SHAPE_RES_TYPE, *COMPUTATION_RES_TYPE, *LITERA
 
 ERL_NIF_TERM ok, bad;
 
-/* These are global instances of the main XLA API. My understanding is that it's correct
- * only to have and maintain one instance of each of these, so I figured it's best to keep them
- * as private data members in the environment. It's convenient not to have to pass references
- * between functions.
- *
- * I think we need to synchronize access to these resources, but I also
- * can't really think of a use case where you'd run in to problems if we didn't.
- */
+// TODO: Synchronize access.
+// TODO: Separate client from global object.
 typedef struct {
   xla::XlaBuilder* builder;
   xla::LocalClient* client;
 } XLA;
 
 // Once the resource is garbage collected, this should also deallocate the C++ object.
+// TODO: address memory leaks
 void free_op(ErlNifEnv* env, void* obj){return;}
 void free_shape(ErlNifEnv* env, void* obj){return;}
 void free_computation(ErlNifEnv* env, void* obj){return;}
@@ -38,25 +33,16 @@ void free_local_executable(ErlNifEnv* env, void* obj){return;}
 void free_shaped_buffer(ErlNifEnv* env, void* obj){return;}
 
 static int open_resources(ErlNifEnv* env) {
-  const char* mod = "XLA";
-  const char* name_op = "Op";
-  const char* name_shape = "Shape";
-  const char* name_computation = "Computation";
-  const char* name_literal = "Literal";
-  const char* name_local_executable = "LocalExecutable";
-  const char* name_shaped_buffer = "ShapedBuffer";
 
-  int flags = ERL_NIF_RT_CREATE | ERL_NIF_RT_TAKEOVER;
+  const char* mod = "EXLA";
 
-  OP_RES_TYPE = enif_open_resource_type(env, mod, name_op, free_op, (ErlNifResourceFlags) flags, NULL);
-  SHAPE_RES_TYPE = enif_open_resource_type(env, mod, name_shape, free_shape, (ErlNifResourceFlags) flags, NULL);
-  COMPUTATION_RES_TYPE = enif_open_resource_type(env, mod, name_computation, free_computation, (ErlNifResourceFlags) flags, NULL);
-  LITERAL_RES_TYPE = enif_open_resource_type(env, mod, name_literal, free_literal, (ErlNifResourceFlags) flags, NULL);
-  LOCAL_EXECUTABLE_RES_TYPE = enif_open_resource_type(env, mod, name_local_executable, free_local_executable, (ErlNifResourceFlags) flags, NULL);
-  SHAPED_BUFFER_RES_TYPE = enif_open_resource_type(env, mod, name_shaped_buffer, free_shaped_buffer, (ErlNifResourceFlags) flags, NULL);
-
-  if(OP_RES_TYPE == NULL || SHAPE_RES_TYPE == NULL || COMPUTATION_RES_TYPE == NULL || LITERAL_RES_TYPE == NULL || LOCAL_EXECUTABLE_RES_TYPE == NULL || SHAPED_BUFFER_RES_TYPE == NULL) return -1;
-  return 0;
+  ExlaNifUtil::open_resource<xla::XlaOp>(env, mod, "Op", free_op);
+  ExlaNifUtil::open_resource<xla::Shape>(env, mod, "Shape", free_shape);
+  ExlaNifUtil::open_resource<xla::XlaComputation>(env, mod, "Computation", free_computation);
+  ExlaNifUtil::open_resource<xla::Literal>(env, mod, "Literal", free_literal);
+  ExlaNifUtil::open_resource<xla::LocalExecutable>(env, mod, "LocalExecutable", free_local_executable);
+  ExlaNifUtil::open_resource<xla::ShapedBuffer>(env, mod, "ShapedBuffer", free_shaped_buffer);
+  return 1;
 }
 
 static int load(ErlNifEnv* env, void** priv, ERL_NIF_TERM load_info){
@@ -67,8 +53,8 @@ static int load(ErlNifEnv* env, void** priv, ERL_NIF_TERM load_info){
 
   XLA* xla_objects;
   xla_objects = (XLA*) enif_alloc(sizeof(XLA));
+
   xla_objects->builder = new xla::XlaBuilder("Elixir");
-  // Client needs to be set.
   xla_objects->client = NULL;
 
   *priv = (void*) xla_objects;
@@ -76,183 +62,7 @@ static int load(ErlNifEnv* env, void** priv, ERL_NIF_TERM load_info){
   return 0;
 }
 
-// TODO: Handle statusor gracefully.
-int enif_get_platform(ErlNifEnv* env, ERL_NIF_TERM term, stream_executor::Platform* platform){
-  std::string platform_str;
-
-  if(!ExlaNifUtil::get_atom(env, term, platform_str)){
-    return 0;
-  }
-
-  if(platform_str.compare("host") == 0){
-    platform = xla::PlatformUtil::GetPlatform("Host").ConsumeValueOrDie();
-    return 1;
-  } else if(platform_str.compare("cuda") == 0){
-    platform = xla::PlatformUtil::GetPlatform("Cuda").ConsumeValueOrDie();
-    return 1;
-  } else {
-    return 0;
-  }
-}
-
-// TODO: XLA has a function for this, but it accepts the shorter type
-int enif_get_primitive_type(ErlNifEnv* env, ERL_NIF_TERM term, xla::PrimitiveType &type){
-  std::string type_str;
-
-  if(!ExlaNifUtil::get_atom(env, term, type_str)) {
-    std::cout << type_str << std::endl;
-    type = xla::PrimitiveType::PRIMITIVE_TYPE_INVALID;
-    return 1;
-  }
-
-  if(type_str.compare("pred") == 0){
-    type = xla::PrimitiveType::PRED;
-  } else if(type_str.compare("int8") == 0){
-    type = xla::PrimitiveType::S8;
-  } else if(type_str.compare("int16") == 0){
-    type = xla::PrimitiveType::S16;
-  } else if(type_str.compare("int32") == 0){
-    type = xla::PrimitiveType::S32;
-  } else if(type_str.compare("int64") == 0){
-    type = xla::PrimitiveType::S64;
-  } else if(type_str.compare("uint8") == 0){
-    type = xla::PrimitiveType::U8;
-  } else if(type_str.compare("uint16") == 0){
-    type = xla::PrimitiveType::U16;
-  } else if(type_str.compare("uint32") == 0){
-    type = xla::PrimitiveType::U32;
-  } else if(type_str.compare("uint64") == 0){
-    type = xla::PrimitiveType::U64;
-  } else if(type_str.compare("float16") == 0){
-    type = xla::PrimitiveType::F16;
-  } else if(type_str.compare("bfloat16") == 0){
-    type = xla::PrimitiveType::BF16;
-  } else if(type_str.compare("float32") == 0){
-    type = xla::PrimitiveType::F32;
-  } else if(type_str.compare("float64") == 0){
-    type = xla::PrimitiveType::F64;
-  } else if(type_str.compare("complex64") == 0){
-    type = xla::PrimitiveType::C64;
-  } else if(type_str.compare("complex128") == 0){
-    type = xla::PrimitiveType::C128;
-  } else if(type_str.compare("tuple") == 0){
-    type = xla::PrimitiveType::TUPLE;
-  } else if(type_str.compare("opaque") == 0){
-    type = xla::PrimitiveType::OPAQUE_TYPE;
-  } else if(type_str.compare("token") == 0){
-    type = xla::PrimitiveType::TOKEN;
-  } else {
-    type = xla::PrimitiveType::PRIMITIVE_TYPE_INVALID;
-  }
-  return 1;
-}
-
-ERL_NIF_TERM enif_make_op(ErlNifEnv* env, xla::XlaOp value){
-  void* ptr = enif_alloc_resource(OP_RES_TYPE, sizeof(xla::XlaOp));
-  new(ptr) xla::XlaOp(std::move(value));
-  ERL_NIF_TERM ret = enif_make_resource(env, ptr);
-  return ret;
-}
-
-ERL_NIF_TERM enif_make_shape(ErlNifEnv* env, xla::Shape value){
-  void* ptr = enif_alloc_resource(SHAPE_RES_TYPE, sizeof(xla::Shape));
-  new(ptr) xla::Shape(std::move(value));
-  ERL_NIF_TERM ret = enif_make_resource(env, ptr);
-  return ret;
-}
-
-ERL_NIF_TERM enif_make_computation(ErlNifEnv* env, xla::XlaComputation value){
-  void* ptr = enif_alloc_resource(COMPUTATION_RES_TYPE, sizeof(xla::XlaComputation));
-  new(ptr) xla::XlaComputation(std::move(value));
-  ERL_NIF_TERM ret = enif_make_resource(env, ptr);
-  return ret;
-}
-
-ERL_NIF_TERM enif_make_literal(ErlNifEnv* env, xla::Literal& value){
-  void* ptr = enif_alloc_resource(LITERAL_RES_TYPE, sizeof(xla::Literal));
-  new(ptr) xla::Literal(std::move(value));
-  ERL_NIF_TERM ret = enif_make_resource(env, ptr);
-  return ret;
-}
-
-ERL_NIF_TERM enif_make_local_executable(ErlNifEnv* env, std::unique_ptr<xla::LocalExecutable>& value){
-  void* ptr = enif_alloc_resource(LOCAL_EXECUTABLE_RES_TYPE, sizeof(xla::LocalExecutable));
-  xla::LocalExecutable* exec = value.release();
-  new(ptr) xla::LocalExecutable(std::move(*exec));
-  return enif_make_resource(env, ptr);
-}
-
-ERL_NIF_TERM enif_make_shaped_buffer(ErlNifEnv* env, xla::ShapedBuffer& value){
-  void* ptr = enif_alloc_resource(SHAPED_BUFFER_RES_TYPE, sizeof(xla::ShapedBuffer));
-  new(ptr) xla::ShapedBuffer(std::move(value));
-  ERL_NIF_TERM ret = enif_make_resource(env, ptr);
-  return ret;
-}
-
-// TODO: Template this.
-// TODO: This should return an integer status instead of the span and rather accept a reference to the Span.
-// TODO: We need to exit gracefully on badarg.
-template <typename T>
-int enif_get_span(ErlNifEnv* env, ERL_NIF_TERM tuple, absl::Span<T> &span){
-  const ERL_NIF_TERM* elems;
-  int num_elems;
-  if(!enif_get_tuple(env, tuple, &num_elems, &elems)) return 0;
-  T data[num_elems];
-  for(int i=0;i<num_elems;i++){
-    T elem;
-    if(!ExlaNifUtil::get(env, elems[i], elem)) return 0;
-    data[i] = elem;
-  }
-  span = absl::Span<T>(data, num_elems);
-  return 1;
-}
-
-// TODO: Template this with above!
-absl::Span<xla::ShapedBuffer*> enif_get_arguments(ErlNifEnv* env, ERL_NIF_TERM tuple){
-  const ERL_NIF_TERM* args;
-  int num_args;
-  enif_get_tuple(env, tuple, &num_args, &args);
-  xla::ShapedBuffer* arguments[num_args];
-  for(int i=0;i<num_args;i++){
-    xla::ShapedBuffer* buffer;
-    enif_get_resource(env, args[i], SHAPED_BUFFER_RES_TYPE, (void **) &buffer);
-    LOG(ERROR) << "buffer on_host_shape: " << buffer->on_host_shape().DebugString();
-    LOG(ERROR) << "buffer on_device_shape: " << buffer->on_device_shape().DebugString();
-    arguments[i] = buffer;
-  }
-  return absl::Span<xla::ShapedBuffer*>(arguments, num_args);
-}
-
-// TODO: Template this with above!
-absl::Span<xla::Shape*> enif_get_argument_layouts(ErlNifEnv* env, ERL_NIF_TERM tuple){
-  const ERL_NIF_TERM* arg_layouts;
-  int num_arg_layouts;
-  enif_get_tuple(env, tuple, &num_arg_layouts, &arg_layouts);
-  xla::Shape* argument_layouts[num_arg_layouts];
-  for(int i=0;i<num_arg_layouts;i++){
-    xla::Shape* shape;
-    enif_get_resource(env, arg_layouts[i], SHAPE_RES_TYPE, (void **) &shape);
-    argument_layouts[i] = shape;
-  }
-  return absl::Span<xla::Shape*>(argument_layouts, num_arg_layouts);
-}
-
-xla::ExecutableBuildOptions enif_get_executable_build_options(ErlNifEnv* env, ERL_NIF_TERM build_options){
-  return xla::ExecutableBuildOptions();
-}
-
-xla::ExecutableRunOptions& enif_get_executable_run_options(ErlNifEnv* env, ERL_NIF_TERM options){
-  XLA* xla_objects = (XLA*) enif_priv_data(env);
-  xla::ExecutableRunOptions* run_options = new xla::ExecutableRunOptions();
-  auto* allocator = new xla::ExlaAllocator(xla_objects->client->platform());
-  run_options->set_allocator(allocator);
-  return *run_options;
-}
 /************************ xla::ShapedBuffer Functions *********************/
-/*
- * Because of what this function actually does, it makes sense to name it to something more idiomatic like:
- * `place_on_device` or `to_device`.
- */
 ERL_NIF_TERM binary_to_shaped_buffer(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   XLA* xla_objects = (XLA*) enif_priv_data(env);
 
@@ -260,7 +70,7 @@ ERL_NIF_TERM binary_to_shaped_buffer(ErlNifEnv* env, int argc, const ERL_NIF_TER
   enif_inspect_binary(env, argv[0], &bin);
 
   xla::Shape* shape;
-  enif_get_resource(env, argv[1], SHAPE_RES_TYPE, (void **) &shape);
+  ExlaNifUtil::get<xla::Shape>(env, argv[1], shape);
 
   const char *data_ptr = (char *) bin.data;
   int64_t data_size = bin.size;
@@ -275,7 +85,7 @@ ERL_NIF_TERM binary_to_shaped_buffer(ErlNifEnv* env, int argc, const ERL_NIF_TER
 
   inp = buffer.release();
 
-  return enif_make_shaped_buffer(env, *inp);
+  return ExlaNifUtil::make<xla::ShapedBuffer>(env, *inp);
 }
 
 /************************ xla::Shape Functions ***************************/
@@ -283,11 +93,11 @@ ERL_NIF_TERM make_shape(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   xla::PrimitiveType element_type;
   absl::Span<long long int> dims;
 
-  if(!enif_get_primitive_type(env, argv[0], element_type)) return enif_make_badarg(env);
-  if(!enif_get_span(env, argv[1], dims)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_type(env, argv[0], element_type)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_span<long long int>(env, argv[1], dims)) return enif_make_badarg(env);
 
   xla::Shape shape = xla::ShapeUtil::MakeShape(element_type, dims);
-  return enif_make_shape(env, shape);
+  return ExlaNifUtil::make<xla::Shape>(env, shape);
 }
 ERL_NIF_TERM make_scalar_shape(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   if(argc != 1){
@@ -295,10 +105,10 @@ ERL_NIF_TERM make_scalar_shape(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv
   }
 
   xla::PrimitiveType element_type;
-  if(!enif_get_primitive_type(env, argv[0], element_type)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_type(env, argv[0], element_type)) return enif_make_badarg(env);
 
   xla::Shape shape = xla::ShapeUtil::MakeScalarShape(element_type);
-  return enif_make_shape(env, shape);
+  return ExlaNifUtil::make<xla::Shape>(env, shape);
 }
 
 ERL_NIF_TERM human_string(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
@@ -307,9 +117,10 @@ ERL_NIF_TERM human_string(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   }
 
   xla::Shape* shape;
-  if(!enif_get_resource(env, argv[0], SHAPE_RES_TYPE, (void **) &shape)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get<xla::Shape>(env, argv[0], shape)) return enif_make_badarg(env);
+
   std::string result = xla::ShapeUtil::HumanString(*shape);
-  return enif_make_string(env, result.c_str(), ERL_NIF_LATIN1);
+  return ExlaNifUtil::make(env, result);
 }
 
 /*********************** xla::LiteralUtil Functions *************************/
@@ -319,21 +130,22 @@ ERL_NIF_TERM create_r0(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   }
   // TODO: Handle other types.
   int data;
-  enif_get_int(env, argv[0], &data);
+
+  if(!ExlaNifUtil::get(env, argv[0], data)) return enif_make_badarg(env);
+
   xla::Literal literal = xla::LiteralUtil::CreateR0(data);
-  return enif_make_literal(env, literal);
+  return ExlaNifUtil::make<xla::Literal>(env, literal);
 }
 
-// This is mainly for testing purposes
 ERL_NIF_TERM literal_to_string(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   if(argc != 1){
     return enif_make_badarg(env);
   }
 
   xla::Literal* literal;
-  enif_get_resource(env, argv[0], LITERAL_RES_TYPE, (void **) &literal);
+  ExlaNifUtil::get<xla::Literal>(env, argv[0], literal);
   std::string literal_str = (*literal).ToString();
-  return enif_make_string(env, literal_str.c_str(), ERL_NIF_LATIN1);
+  return ExlaNifUtil::make(env, literal_str);
 }
 
 /************************ xla::XlaOp Functions ***************************/
@@ -348,15 +160,14 @@ ERL_NIF_TERM parameter(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   xla::Shape* shape;
   std::string name;
 
-  if(!enif_get_int64(env, argv[0], &param_num)) return enif_make_badarg(env);
-  if(!enif_get_resource(env, argv[1], SHAPE_RES_TYPE, (void **) &shape)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get(env, argv[0], param_num)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get<xla::Shape>(env, argv[1], shape)) return enif_make_badarg(env);
   if(!ExlaNifUtil::get(env, argv[2], name)) return enif_make_badarg(env);
 
   xla::XlaOp op = xla::Parameter(xla_objects->builder, param_num, *shape, name);
-  return enif_make_op(env, op);
+  return ExlaNifUtil::make<xla::XlaOp>(env, op);
 }
 
-/* Stub for element-wise binary functions */
 ERL_NIF_TERM xla_binary_op(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[], xla::XlaOp(*lambda)(xla::XlaOp, xla::XlaOp, absl::Span<const long long int>)){
   if(argc != 3){
     return enif_make_badarg(env);
@@ -364,12 +175,13 @@ ERL_NIF_TERM xla_binary_op(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[], 
 
   xla::XlaOp *lhs, *rhs;
   absl::Span<long long int> broadcast_dims;
-  if(!enif_get_resource(env, argv[0], OP_RES_TYPE, (void **) &lhs)) return enif_make_badarg(env);
-  if(!enif_get_resource(env, argv[1], OP_RES_TYPE, (void **) &rhs)) return enif_make_badarg(env);
-  if(!enif_get_span(env, argv[2], broadcast_dims)) return enif_make_badarg(env);
+
+  if(!ExlaNifUtil::get<xla::XlaOp>(env, argv[0], lhs)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get<xla::XlaOp>(env, argv[1], rhs)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_span(env, argv[2], broadcast_dims)) return enif_make_badarg(env);
 
   xla::XlaOp result = lambda(*lhs, *rhs, broadcast_dims);
-  return enif_make_op(env, result);
+  return ExlaNifUtil::make<xla::XlaOp>(env, result);
 }
 
 ERL_NIF_TERM add(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla_binary_op(env, argc, argv, xla::Add);}
@@ -401,17 +213,17 @@ ERL_NIF_TERM pow(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla
 ERL_NIF_TERM complex(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla_binary_op(env, argc, argv, xla::Complex);}
 ERL_NIF_TERM atan2(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla_binary_op(env, argc, argv, xla::Atan2);}
 
-/* Stub for element-wise unary functions */
 ERL_NIF_TERM xla_unary_op(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[], xla::XlaOp(*lambda)(xla::XlaOp)){
   if(argc != 1){
     return enif_make_badarg(env);
   }
 
   xla::XlaOp *op;
-  if(!enif_get_resource(env, argv[0], OP_RES_TYPE, (void **) &op)) return enif_make_badarg(env);
+
+  if(!ExlaNifUtil::get<xla::XlaOp>(env, argv[0], op)) return enif_make_badarg(env);
 
   xla::XlaOp result = lambda(*op);
-  return enif_make_op(env, result);
+  return ExlaNifUtil::make<xla::XlaOp>(env, result);
 }
 
 ERL_NIF_TERM abs(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla_unary_op(env, argc, argv, xla::Abs);}
@@ -440,17 +252,19 @@ ERL_NIF_TERM conj(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xl
 ERL_NIF_TERM population_count(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla_unary_op(env, argc, argv, xla::PopulationCount);}
 ERL_NIF_TERM copy(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){return xla_unary_op(env, argc, argv, xla::Copy);}
 
-// Constant Creation Methods
 ERL_NIF_TERM constant_r0(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   if(argc != 1){
     return enif_make_badarg(env);
   }
 
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   int value;
-  enif_get_int(env, argv[0], &value);
+
+  if(!ExlaNifUtil::get(env, argv[0], value)) return enif_make_badarg(env);
+
   xla::XlaOp op = xla::ConstantR0(xla_objects->builder, value);
-  return enif_make_op(env, op);
+  return ExlaNifUtil::make<xla::XlaOp>(env, op);
 }
 
 ERL_NIF_TERM constant_r1_fill(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
@@ -459,38 +273,43 @@ ERL_NIF_TERM constant_r1_fill(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[
   }
 
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   int length, value;
-  enif_get_int(env, argv[0], &length);
-  enif_get_int(env, argv[1], &value);
+
+  if(!ExlaNifUtil::get(env, argv[0], length)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get(env, argv[1], value)) return enif_make_badarg(env);
+
   xla::XlaOp op = xla::ConstantR1(xla_objects->builder, length, value);
-  return enif_make_op(env, op);
+  return ExlaNifUtil::make<xla::XlaOp>(env, op);
 }
 
 ERL_NIF_TERM dot(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 2){
+    return enif_make_badarg(env);
+  }
+
   xla::XlaOp *lhs, *rhs;
-  enif_get_resource(env, argv[0], OP_RES_TYPE, (void **) &lhs);
-  enif_get_resource(env, argv[1], OP_RES_TYPE, (void **) &rhs);
+
+  if(!ExlaNifUtil::get<xla::XlaOp>(env, argv[0], lhs)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get<xla::XlaOp>(env, argv[1], rhs)) return enif_make_badarg(env);
   // TODO: Handle Precision Configuration
   xla::XlaOp result = xla::Dot(*lhs, *rhs);
-  return enif_make_op(env, result);
+  return ExlaNifUtil::make<xla::XlaOp>(env, result);
 }
 
 /************************ xla::ClientLibrary Functions ***************************/
-/*
- * This creates the local client which interfaces with the underlying XLA service.
- * It usually takes config ops, but I haven't handled those yet.
- */
 ERL_NIF_TERM get_or_create_local_client(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   if(argc != 1){
     return enif_make_badarg(env);
   }
 
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   stream_executor::Platform* platform;
 
-  if(!enif_get_platform(env, argv[0], platform)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_platform(env, argv[0], platform)) return enif_make_badarg(env);
 
-  // Todo: Handle this gracefully
+  // TODO: Handle this gracefully
   xla::StatusOr<xla::LocalClient*> client_status = xla::ClientLibrary::GetOrCreateLocalClient(platform);
   // This matches really nicely with the ! pattern
   xla::LocalClient* client = client_status.ConsumeValueOrDie();
@@ -499,9 +318,14 @@ ERL_NIF_TERM get_or_create_local_client(ErlNifEnv* env, int argc, const ERL_NIF_
 }
 
 ERL_NIF_TERM get_device_count(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 1){
+    return enif_make_badarg(env);
+  }
+
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   if(xla_objects->client == NULL){
-    return enif_make_tuple2(env, bad, enif_make_string(env, "No client available. Try creating one with get_or_create_local_client/0.", ERL_NIF_LATIN1));
+    return ExlaNifUtil::error(env, "No client found.");
   }
 
   int device_count = xla_objects->client->device_count();
@@ -510,27 +334,40 @@ ERL_NIF_TERM get_device_count(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[
 
 /************ Build, Compilation, Execution *************/
 ERL_NIF_TERM build(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 1){
+    return enif_make_badarg(env);
+  }
+
   XLA* xla_objects = (XLA*) enif_priv_data(env);
 
   xla::XlaOp* root;
-  // TODO: Handle args
-  enif_get_resource(env, argv[0], OP_RES_TYPE, (void **) &root);
+
+  if(!ExlaNifUtil::get<xla::XlaOp>(env, argv[0], root)) return enif_make_badarg(env);
+
   xla::StatusOr<xla::XlaComputation> computation_status = xla_objects->builder->Build(*root);
   // TODO: Handle StatusOr more gracefully.
-  return enif_make_computation(env, computation_status.ConsumeValueOrDie());
+  xla::XlaComputation computation = computation_status.ConsumeValueOrDie();
+  return ExlaNifUtil::make<xla::XlaComputation>(env, computation);
 }
 
 ERL_NIF_TERM compile(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 3){
+    return enif_make_badarg(env);
+  }
+
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   if(xla_objects->client == NULL){
-    return enif_make_tuple2(env, bad, enif_make_string(env, "No client found. Try creating one with get_or_create_local_client/0.", ERL_NIF_LATIN1));
+    return ExlaNifUtil::error(env, "No client found.");
   }
 
   xla::XlaComputation* computation;
-  // TODO: Handle Args
-  enif_get_resource(env, argv[0], COMPUTATION_RES_TYPE, (void **) &computation);
-  absl::Span<xla::Shape*> argument_layouts = enif_get_argument_layouts(env, argv[1]);
-  xla::ExecutableBuildOptions options = enif_get_executable_build_options(env, argv[2]);
+  absl::Span<xla::Shape*> argument_layouts;
+  xla::ExecutableBuildOptions options;
+
+  if(!ExlaNifUtil::get<xla::XlaComputation>(env, argv[0], computation)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_span(env, argv[1], argument_layouts)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_options(env, argv[2], options)) return enif_make_badarg(env);
 
   xla::StatusOr<std::vector<std::unique_ptr<xla::LocalExecutable>>> exec_status = xla_objects->client->Compile(*computation, argument_layouts, options);
   // TODO: Handle this gracefully.
@@ -538,56 +375,72 @@ ERL_NIF_TERM compile(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
   ERL_NIF_TERM exec_refs[executables.size()];
   int i = 0;
   for(auto it=std::begin(executables);it!=std::end(executables);++it){
-    exec_refs[i++] = enif_make_local_executable(env, executables.at(i));
+    exec_refs[i++] = ExlaNifUtil::make<xla::LocalExecutable>(env, executables.at(i));
   }
   return exec_refs[0];
 }
 
 ERL_NIF_TERM run(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 3){
+    return enif_make_badarg(env);
+  }
+
   xla::LocalExecutable* local_executable;
-  // TODO: Handle Args
-  enif_get_resource(env, argv[0], LOCAL_EXECUTABLE_RES_TYPE, (void **) &local_executable);
-  absl::Span<xla::ShapedBuffer*> arguments = enif_get_arguments(env, argv[1]);
-  xla::ExecutableRunOptions run_options = enif_get_executable_run_options(env, argv[2]);
+  absl::Span<xla::ShapedBuffer*> arguments;
+  xla::ExecutableRunOptions run_options;
+
+  if(!ExlaNifUtil::get<xla::LocalExecutable>(env, argv[0], local_executable)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_span(env, argv[1], arguments)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get_options(env, argv[2], run_options)) return enif_make_badarg(env);
 
   xla::StatusOr<xla::ScopedShapedBuffer> run_status = local_executable->Run(arguments, run_options);
   // TODO: Handle this gracefully
   xla::ScopedShapedBuffer result = run_status.ConsumeValueOrDie();
-  return enif_make_shaped_buffer(env, result);
+  return ExlaNifUtil::make<xla::ShapedBuffer>(env, result);
 }
 
 ERL_NIF_TERM literal_to_shaped_buffer(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 3){
+    return enif_make_badarg(env);
+  }
+
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   if(xla_objects->client == NULL){
-    return enif_make_tuple2(env, bad, enif_make_string(env, "No client found. Try creating one with get_or_create_local_client/0.", ERL_NIF_LATIN1));
+    return ExlaNifUtil::error(env, "No client found.");
   }
 
   xla::Literal* literal;
   int device_ordinal;
-  // TODO: Handle Args
-  enif_get_resource(env, argv[0], LITERAL_RES_TYPE, (void **) &literal);
-  enif_get_int(env, argv[1], &device_ordinal);
+
+  if(!ExlaNifUtil::get<xla::Literal>(env, argv[0], literal)) return enif_make_badarg(env);
+  if(!ExlaNifUtil::get(env, argv[1], device_ordinal)) return enif_make_badarg(env);
 
   xla::StatusOr<xla::ScopedShapedBuffer> buffer_status = xla_objects->client->LiteralToShapedBuffer(*literal, device_ordinal);
   // TODO: Handle this gracefully.
   xla::ScopedShapedBuffer buffer = buffer_status.ConsumeValueOrDie();
-  return enif_make_shaped_buffer(env, buffer);
+  return ExlaNifUtil::make<xla::ShapedBuffer>(env, buffer);
 }
 
 ERL_NIF_TERM shaped_buffer_to_literal(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 1){
+    return enif_make_badarg(env);
+  }
+
   XLA* xla_objects = (XLA*) enif_priv_data(env);
+
   if(xla_objects->client == NULL){
-    return enif_make_tuple2(env, bad, enif_make_string(env, "No client found. Try creating one with get_or_create_local_client/0.", ERL_NIF_LATIN1));
+    return ExlaNifUtil::error(env, "No client found.");
   }
 
   xla::ShapedBuffer* buffer;
-  // TODO: Handle Args
-  enif_get_resource(env, argv[0], SHAPED_BUFFER_RES_TYPE, (void **) &buffer);
+
+  if(!ExlaNifUtil::get<xla::ShapedBuffer>(env, argv[0], buffer)) return enif_make_badarg(env);
 
   xla::StatusOr<xla::Literal> literal_status = xla_objects->client->ShapedBufferToLiteral(*buffer);
   // TODO: Handle this gracefully.
   xla::Literal literal = literal_status.ConsumeValueOrDie();
-  return enif_make_literal(env, literal);
+  return ExlaNifUtil::make<xla::Literal>(env, literal);
 }
 
 /*********** HLO Methods *************/
@@ -600,9 +453,15 @@ xla::StatusOr<std::unique_ptr<xla::HloModule>> get_hlo_module(const xla::XlaComp
 }
 
 ERL_NIF_TERM get_computation_hlo_text(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 1){
+    return enif_make_badarg(env);
+  }
+
   // TODO: Handle this gracefully
   xla::XlaComputation* computation;
-  enif_get_resource(env, argv[0], COMPUTATION_RES_TYPE, (void **) &computation);
+
+  if(!ExlaNifUtil::get<xla::XlaComputation>(env, argv[0], computation)) return enif_make_badarg(env);
+
   xla::StatusOr<std::unique_ptr<xla::HloModule>> hlo_module_status = get_hlo_module(*computation);
   // // TODO: Handle this gracefully
   std::unique_ptr<xla::HloModule> hlo_module = hlo_module_status.ConsumeValueOrDie();
@@ -611,15 +470,21 @@ ERL_NIF_TERM get_computation_hlo_text(ErlNifEnv* env, int argc, const ERL_NIF_TE
   options = xla::HloPrintOptions::ShortParsable();
   options.set_print_large_constants(false);
   std::string result = hlo_module->ToString(options);
-  return enif_make_string(env, result.c_str(), ERL_NIF_LATIN1);
+  return ExlaNifUtil::make(env, result);
 }
 
 ERL_NIF_TERM get_computation_hlo_proto(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]){
+  if(argc != 1){
+    return enif_make_badarg(env);
+  }
+
   xla::XlaComputation* computation;
-  enif_get_resource(env, argv[0], COMPUTATION_RES_TYPE, (void **) &computation);
+
+  if(!ExlaNifUtil::get<xla::XlaComputation>(env, argv[0], computation)) return enif_make_badarg(env);
+
   std::string result;
   (*computation).proto().SerializeToString(&result);
-  return enif_make_string(env, result.c_str(), ERL_NIF_LATIN1);
+  return ExlaNifUtil::make(env, result);
 }
 
 static ErlNifFunc exla_funcs[] = {
@@ -706,4 +571,4 @@ static ErlNifFunc exla_funcs[] = {
   {"get_computation_hlo_text", 1, get_computation_hlo_text}
 };
 
-ERL_NIF_INIT(Elixir.Exla, exla_funcs, &load, NULL, NULL, NULL);
+ERL_NIF_INIT(Elixir.Exla.NIF, exla_funcs, &load, NULL, NULL, NULL);
